ch-6 feat: добавлен компонент Line
