+++ src/app/page.tsx
@@ -1,9 +1,11 @@
-import Header from "@/components/Header/Header";
+import Header from "@/components/Header";
+import Line from "@/components/Line";
 
 export default function Home() {
   return (
    <div>
      <Header/>
+       <Line/>
    </div>
   );
 }
