+++ src/components/Line.tsx
@@ -1,0 +1,7 @@
+
+const Line = () => {
+  return (
+    <hr className="w-full mt-3"></hr>
+  );
+};
+export default Line;
